rename API version from IP_HEADERS to DSCP
DSCP better reflects the changes to the API types/endpoints, so make the
name more appropriate.
