+++ mg-api-types/versions/src/lib.rs
@@ -35,7 +35,7 @@ pub mod latest;
 pub mod v1;
 #[path = "v4_over_v6_static_routes/mod.rs"]
 pub mod v10;
-#[path = "ip_headers/mod.rs"]
+#[path = "dscp/mod.rs"]
 pub mod v11;
 #[path = "ipv6_basic/mod.rs"]
 pub mod v2;
